requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy<2
opencv-python-headless==4.9.0.80
python-doctr[torch]==0.8.1
ultralytics==8.1.24
